Move screening narrative selectors to proper directory
[#22]
