+++ app/javascript/selectors/screening/narrativeShowSelectors.js
